Adjust minHeight of post image depending on image ratio

+++ domain/src/main/kotlin/io/github/jasvilladarez/domain/entity/ImageVersion.kt
@@ -35,9 +35,9 @@ data class ImageVersion(
 
 data class ImageVersionMetaData(
         @SerializedName("width")
-        val width: Long,
+        val width: Int,
         @SerializedName("height")
-        val height: Long,
+        val height: Int,
         @SerializedName("type")
         val mimeType: String
 )
